Rename: 유틸리티 함수 util 폴더로 이동
